% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determineOptimalModel.R
\name{determineOptimalModel}
\alias{determineOptimalModel}
\title{Determine Optimal GMM}
\usage{
determineOptimalModel(bestPCSet)
}
\arguments{
\item{bestPCSet}{a matrix whose columns contain the optimal principal components.}
}
\value{
summary of output from mclust::mclustBIC(<bestPCSet>) computed with optimal model parameters.
}
\description{
Use log likelihood criterion to select optimal model for best PC set.
}
\examples{
data <- validateAndLoadData(iris)
pcObj <- prcomp(data)
pcData <- pcObj$x
iterationResults <- executePCFiltering(pcData)
bestPCSet <- iterationResults[[length(iterationResults)]]
clusterResults <- evaluateClusterQuality(bestPCSet)
optimalModel <- determineOptimalModel(bestPCSet)
}
\seealso{
\code{\link[mclust]{mclustBIC}}
}
