% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateClusterQuality.R
\name{evaluateClusterQuality}
\alias{evaluateClusterQuality}
\title{Evaluate PC Cluster Quality}
\usage{
evaluateClusterQuality(pcMatrix)
}
\arguments{
\item{pcMatrix}{a matrix whose columns contain the principal components.}
}
\value{
list containing optimal model characteristics and classification
}
\description{
compute Bayesian information criterion (BIC) for Gaussian mixture models (GMM) and associated log likelihood to quantify cluster quality for a subset of PCs.
}
\examples{
data <- validateAndLoadData(iris)
pcObj <- prcomp(data)
pcData <- pcObj$x
iterationResults <- executePCFiltering(pcData)
bestPCSet <- iterationResults[[length(iterationResults)]]
clusterResults <- evaluateClusterQuality(bestPCSet)
}
\seealso{
\code{\link[mclust]{mclustBIC}}

\code{\link[mclust]{summary.mclustBIC}}
}
