% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executePCFiltering.R
\name{executePCFiltering}
\alias{executePCFiltering}
\title{Filter principal component data}
\usage{
executePCFiltering(pcData)
}
\arguments{
\item{pcData}{a numerical R matrix or data frame.}
}
\value{
list of iteration results from PCA filtering, where each iteation has one less PC. List elements are PCA matrices. Last entry in the list corresponds to the top 2 PCs for clustering.
}
\description{
Filter raw principal component data automatically until only the top 2 PCs remain.
}
\details{
Indirect recursion in the case of optimal subset ambiguity (i.e. different models for maxBIC and maxLikelihood)
}
\examples{
data <- validateAndLoadData(iris)
pcObj <- prcomp(data)
pcData <- pcObj$x
iterationResults <- executePCFiltering(pcData)
}
\seealso{
\code{\link{evaluateClusterQuality}}
}
