% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNearestNeighbor.R
\name{findNearestNeighbor}
\alias{findNearestNeighbor}
\title{findNearestNeighbor}
\usage{
findNearestNeighbor(pcCol, queryPt)
}
\arguments{
\item{pcCol}{a column from a optimal PC subset}

\item{queryPt}{a number}
}
\value{
value from complete PC matrix closest to query
}
\description{
findNearestNeighbor
}
