% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePCSubsets.R
\name{generatePCSubsets}
\alias{generatePCSubsets}
\title{generatePCSubsets}
\usage{
generatePCSubsets(pcData, subsetSize = NULL)
}
\arguments{
\item{pcData}{a validated R matrix containing numeric scaled data.}

\item{subsetSize}{number of principal components}
}
\value{
list of PCA matrices
}
\description{
generatePCSubsets
}
