% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPcclustApp.R
\name{runPcclustApp}
\alias{runPcclustApp}
\title{\code{runPcclustApp} launch the shiny app distributed with this package framework}
\usage{
runPcclustApp()
}
\description{
\code{runPcclustApp} launches the shiny app for which the code has been placed in  \code{./inst/shiny-scripts/rptApp/}.
}
