% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateAndLoadData.R
\name{validateAndLoadData}
\alias{validateAndLoadData}
\title{Validate and load data}
\usage{
validateAndLoadData(data, isFile = FALSE, scaled = TRUE)
}
\arguments{
\item{data}{a numerical R matrix or data frame.}

\item{isFile}{if TRUE, data accepts a path to a file containing numeric multidimensional data. FALSE by default.}

\item{scaled}{if TRUE, use Z-score standardization so that outliers aren"t weighted less. TRUE by default.}
}
\value{
Centered scaled R matrix with loaded data.
}
\description{
Validate input type and load data into data frame.
}
\details{
Supported input types:
\itemize{
  \item R numerical matrix or data frame
  \item a wide range of file formats (refer to https://github.com/leeper/rio for a full list of supported formats)
}
}
\examples{
irisCSV <- system.file("extdata", "iris.csv", package = "pcclust")
data <- validateAndLoadData(irisCSV, isFile = TRUE)
}
\seealso{
\code{\link[rio]{import}}
}
