% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizePCA.R
\name{visualizePCA}
\alias{visualizePCA}
\title{Generate PCA plots with optimized clustering}
\usage{
visualizePCA(bestPCSet, clusters, pcObj, outDir = ".")
}
\arguments{
\item{bestPCSet}{a matrix whose columns contain the optimal principal components.}

\item{clusters}{vector of predicted labels}

\item{pcObj}{a PCA object}

\item{outDir}{path to ouput directory where pcclust_visualization folder will be generated. Defaults to current working directory.}
}
\value{
path to pcclust_visualization output directory. Outputs high quality .svg files for each of the 3 plots.
}
\description{
Baseline PCA visualizations including standard 2-component PCA plot, density plot, and variance contribution pie chart.
}
\examples{
data <- validateAndLoadData(iris)
pcObj <- prcomp(data)
pcData <- pcObj$x
iterationResults <- executePCFiltering(pcData)
bestPCSet <- iterationResults[[length(iterationResults)]]
clusterResults <- evaluateClusterQuality(bestPCSet)
optimalModel <- determineOptimalModel(bestPCSet)
clusters <- optimalModel$classification
out <- visualizePCA(bestPCSet, clusters, pcObj)
}
\seealso{
\code{\link[ggplot2]{ggplot}}

\code{\link[grDevices]{svg}}
}
