% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoomPC.R
\name{zoomPC}
\alias{zoomPC}
\title{Sample-specific PCA visualizations}
\usage{
zoomPC(x, y, pcData, bestPCSet, clusters, iterationResults, outPath)
}
\arguments{
\item{x}{approximate x coordinate value from optimal PCA plot. See details for more information.}

\item{y}{same as x.}

\item{pcData}{a validated R matrix containing numeric scaled PCA data.}

\item{bestPCSet}{a matrix whose columns contain the optimal principal components.}

\item{clusters}{vector of predicted labels}

\item{iterationResults}{list of iteration results from PCA filtering, where each iteation has one less PC. List elements are PCA matrices. Last entry in the list corresponds to the top 2 PCs for clustering.}

\item{outPath}{path to ouput directory where pcclust_visualization folder was generated.}
}
\value{
NULL. Outputs high quality .svg files in pcclust_visualization for each of the 2 plots.
}
\description{
Supporting PCA visualizations and sample-specific metadata including standard 2-component PCA plot labeled with query, a diverging lollipop chart showing the PC breakdown for the query, and an interactive PCA plot showing the location of the query in 3D space.
}
\examples{
data <- validateAndLoadData(iris)
pcObj <- prcomp(data)
pcData <- pcObj$x
iterationResults <- executePCFiltering(pcData)
bestPCSet <- iterationResults[[length(iterationResults)]]
clusterResults <- evaluateClusterQuality(bestPCSet)
optimalModel <- determineOptimalModel(bestPCSet)
clusters <- optimalModel$classification
out <- visualizePCA(bestPCSet, clusters, pcObj)
x <- 0.5
y <- -0.3
zoomPC(x, y, pcData, bestPCSet, clusters, iterationResults, out)
}
\seealso{
\code{\link[ggplot2]{ggplot}}

\code{\link[grDevices]{svg}}

\code{\link[rgl]{plot3d}}

\code{\link[rgl]{text3d}}
}
